TSC frequency: allow a 2kHz variance when starting/migrating a TSC-enabled VMI
Some nodes on an homogeneous cluster can have a 1 or 2 kHz difference between each other,
as the value is estimated by the kernel at boot time, with a small margin for error.

A 2 kHz variance is only a fraction of the allowable 533kHz, but should fix the vast majority of cases.

Signed-off-by: Jed Lejosne <jed@redhat.com>