migration controller: add a secondary queue to keep the main one managable
When starting a large number of migrations, most of them will stay pending for a long time.
Since pending migrations keep consuming controller execution cycles to just get re-enqueued,
there can be not enough cycles left for active migrations to move forward in time.

This commit introduces a second queue dedicated to the pending migrations.
This second queue will only be consumed when the main one is empty.
That's a bit harsh, and in extreme scenarios could lead to fewer than 5 migrations running
in parallel, but it's better than spinlocking virt-controller doing pointless re-enqueues.

Signed-off-by: Jed Lejosne <jed@redhat.com>