Let target node detect a migration end
This patch allows the target node to detect that a
migration has completed using libvirt events.

This is especially needed on KubeVirt updates as
the cluster will run with an updated handler and
likely some old launchers, without this the migration
will never end as the old handlers were previously
(and wrongly) setting the migration as completed
as soon as a domain would show up, now it's the launcher
that tells the handler when the migration is actually done.

To prevent this we need to target to mark the migration as
completed.

Signed-off-by: Antonio Cardace <acardace@redhat.com>